Completed Version V1

+++ Cloud_Classroom/Cloud_Classroom/CCMiscHelper.m
@@ -13,7 +13,7 @@
 
 +(void)showConnectionFailedAlertWithSendResult:(SendMessageResult)sentResult{
     
-    NSLog(@"Conneciotn failed, send result code: %ld", sentResult);
+    NSLog(@"Conneciotn failed, send result code: %d", (int)sentResult);
     
     [CCMiscHelper showAlertWithTitle:@"Connection failed"
                           andMessage:@"Couldn't connect to server. Please check your network availability."];
